backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
python-dotenv==1.0.0
